fix #7: 리뷰 평점 데이터타입 변경

+++ src/main/java/com/sparta/gourmate/domain/store/dto/AvgResponseDto.java
@@ -7,9 +7,9 @@ import java.util.UUID;
 @Getter
 public class AvgResponseDto {
     private final UUID storeId;
-    private final float avg;
+    private final double avg;
 
-    public AvgResponseDto(UUID storeId, float avg) {
+    public AvgResponseDto(UUID storeId, double avg) {
         this.storeId = storeId;
         this.avg = avg;
     }
